fix(method coverage manafa)

+++ anadroid/application/Application.py
@@ -74,7 +74,14 @@ class App(AbstractApplication):
         all_m = os.path.join(self.local_res, "all", "allMethods.json")
         if not os.path.exists(all_m):
             all_m_proj = os.path.join(self.proj.proj_dir, "allMethods.json")
-            shutil.copyfile(all_m_proj, all_m)
+            if os.path.exists(all_m_proj):
+                print(f"copying {all_m_proj} to {all_m}")
+                shutil.copyfile(all_m_proj, all_m)
+            else:
+                other_possible_all_m = os.path.join(self.local_res, "oldRuns", "all", "allMethods.json")
+                if os.path.exists(other_possible_all_m):
+                    print(f"copying {other_possible_all_m} to {all_m}")
+                    shutil.copyfile(other_possible_all_m, all_m)
 
     def init_local_test_(self, testing_framework, inst_type):
         """initialize directory for a current test being run with testing_framework, with app instrumented with inst_type
